@@ -7,13 +7,13 @@
     "prepublish": "npm run lib",
     "lib": "npm run babel",
     "babel": "rimraf lib && babel src/ -d lib/",
-    "test": "./node_modules/.bin/karma start karma.conf.js",
-    "test-travis": "./node_modules/karma/bin/karma start --browsers Firefox --single-run",
+    "test": "karma start karma.conf.js",
+    "test-travis": "karma start --browsers Firefox --single-run",
     "example": "webpack-dev-server --content-base ./example/ --config ./example/webpack.config.js"
   },
   "peerDependencies": {
-    "react": "^0.14 || ^15.0",
-    "react-dom": "^0.14 || ^15.0"
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4"
   },
   "repository": {
     "type": "git",
@@ -60,9 +60,8 @@
     "karma-webpack": "^1.7.0",
     "mocha": "^2.3.3",
     "node-libs-browser": "^0.5.3",
-    "react": "^0.14.7",
-    "react-addons-test-utils": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-hot-loader": "^1.3.0",
     "rimraf": "^2.4.4",
     "sinon": "uberVU/Sinon.JS.git",
@@ -73,8 +72,9 @@
   "dependencies": {
     "babel-runtime": "^5.8.29",
     "deep-equal": "^1.0.1",
-    "radium": "^0.18.0",
+    "prop-types": "^15.5.8",
+    "radium": "^0.19.0",
     "shallowequal": "^0.2.2",
-    "velocity-react": "^1.1.2"
+    "velocity-react": "^1.3.1"
   }
 }
