@@ -1,108 +1,100 @@
 'use strict';
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import Radium from 'radium';
 import {VelocityComponent} from 'velocity-react';
 
-const Loading = (props) => {
-    return (
-        <div style={props.style}>
-            loading...
-        </div>
-    );
+const Loading = ({style}) => {
+    return <div style={style}>loading...</div>;
 };
-
 Loading.propTypes = {
-    style: React.PropTypes.object
+    style: PropTypes.object
 };
 
-const Toggle = (props) => {
-    const style = props.style;
-    const height = style.height;
-    const width = style.width;
-    let midHeight = height * 0.5;
-    let points = `0,0 0,${height} ${width},${midHeight}`;
+const Toggle = ({style}) => {
+    const {height, width} = style;
+    const midHeight = height * 0.5;
+    const points = `0,0 0,${height} ${width},${midHeight}`;
+
     return (
         <div style={style.base}>
             <div style={style.wrapper}>
                 <svg height={height} width={width}>
-                    <polygon
-                        points={points}
-                        style={style.arrow}
-                    />
+                    <polygon points={points}
+                             style={style.arrow}/>
                 </svg>
             </div>
         </div>
     );
 };
-
 Toggle.propTypes = {
-    style: React.PropTypes.object
+    style: PropTypes.object
 };
 
-const Header = (props) => {
-    const style = props.style;
+const Header = ({node, style}) => {
     return (
         <div style={style.base}>
             <div style={style.title}>
-                {props.node.name}
+                {node.name}
             </div>
         </div>
     );
 };
-
 Header.propTypes = {
-    style: React.PropTypes.object,
-    node: React.PropTypes.object.isRequired
+    style: PropTypes.object,
+    node: PropTypes.object.isRequired
 };
 
 @Radium
 class Container extends React.Component {
-    constructor(props){
-        super(props);
-    }
-    render(){
+    render() {
         const {style, decorators, terminal, onClick, node} = this.props;
+
         return (
-            <div
-                ref="clickable"
-                onClick={onClick}
-                style={style.container}>
-                { !terminal ? this.renderToggle() : null }
-                <decorators.Header
-                    node={node}
-                    style={style.header}
-                />
+            <div onClick={onClick}
+                 ref={ref => this.clickableRef = ref}
+                 style={style.container}>
+                {!terminal ? this.renderToggle() : null}
+
+                <decorators.Header node={node}
+                                   style={style.header}/>
             </div>
         );
     }
-    renderToggle(){
-        const animations = this.props.animations;
-        if(!animations){ return this.renderToggleDecorator(); }
+
+    renderToggle() {
+        const {animations} = this.props;
+
+        if (!animations) {
+            return this.renderToggleDecorator();
+        }
+
         return (
-            <VelocityComponent ref="velocity"
-                duration={animations.toggle.duration}
-                animation={animations.toggle.animation}>
+            <VelocityComponent animation={animations.toggle.animation}
+                               duration={animations.toggle.duration}
+                               ref={ref => this.velocityRef = ref}>
                 {this.renderToggleDecorator()}
             </VelocityComponent>
         );
     }
-    renderToggleDecorator(){
+
+    renderToggleDecorator() {
         const {style, decorators} = this.props;
-        return (<decorators.Toggle style={style.toggle}/>);
+
+        return <decorators.Toggle style={style.toggle}/>;
     }
 }
-
 Container.propTypes = {
-    style: React.PropTypes.object.isRequired,
-    decorators: React.PropTypes.object.isRequired,
-    terminal: React.PropTypes.bool.isRequired,
-    onClick: React.PropTypes.func.isRequired,
-    animations: React.PropTypes.oneOfType([
-        React.PropTypes.object,
-        React.PropTypes.bool
+    style: PropTypes.object.isRequired,
+    decorators: PropTypes.object.isRequired,
+    terminal: PropTypes.bool.isRequired,
+    onClick: PropTypes.func.isRequired,
+    animations: PropTypes.oneOfType([
+        PropTypes.object,
+        PropTypes.bool
     ]).isRequired,
-    node: React.PropTypes.object.isRequired
+    node: PropTypes.object.isRequired
 };
 
 export default {
