@@ -1,6 +1,7 @@
 'use strict';
 
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import TreeNode from './node';
 import defaultDecorators from './decorators';
@@ -8,24 +9,24 @@ import defaultTheme from '../themes/default';
 import defaultAnimations from '../themes/animations';
 
 class TreeBeard extends React.Component {
-    constructor(props){
-        super(props);
-    }
-    render(){
-        let data = this.props.data;
+    render() {
+        const {animations, decorators, data: propsData, onToggle, style} = this.props;
+        let data = propsData;
+
         // Support Multiple Root Nodes. Its not formally a tree, but its a use-case.
-        if(!Array.isArray(data)){ data = [data]; }
+        if (!Array.isArray(data)) {
+            data = [data];
+        }
         return (
-            <ul style={this.props.style.tree.base} ref="treeBase">
+            <ul style={style.tree.base}
+                ref={ref => this.treeBaseRef = ref}>
                 {data.map((node, index) =>
-                    <TreeNode
-                        key={node.id || index}
-                        node={node}
-                        onToggle={this.props.onToggle}
-                        animations={this.props.animations}
-                        decorators={this.props.decorators}
-                        style={this.props.style.tree.node}
-                    />
+                    <TreeNode animations={animations}
+                              decorators={decorators}
+                              key={node.id || index}
+                              node={node}
+                              onToggle={onToggle}
+                              style={style.tree.node}/>
                 )}
             </ul>
         );
@@ -33,17 +34,17 @@ class TreeBeard extends React.Component {
 }
 
 TreeBeard.propTypes = {
-    style: React.PropTypes.object,
-    data: React.PropTypes.oneOfType([
-        React.PropTypes.object,
-        React.PropTypes.array
+    style: PropTypes.object,
+    data: PropTypes.oneOfType([
+        PropTypes.object,
+        PropTypes.array
     ]).isRequired,
-    animations: React.PropTypes.oneOfType([
-        React.PropTypes.object,
-        React.PropTypes.bool
+    animations: PropTypes.oneOfType([
+        PropTypes.object,
+        PropTypes.bool
     ]),
-    onToggle: React.PropTypes.func,
-    decorators: React.PropTypes.object
+    onToggle: PropTypes.func,
+    decorators: PropTypes.object
 };
 
 TreeBeard.defaultProps = {
