@@ -1,22 +1,18 @@
 'use strict';
 
 export default {
-    toggle: (props) => {
-        return {
-            animation: { rotateZ: props.node.toggled ? 90 : 0 },
+    toggle: ({node: {toggled}}) => ({
+        animation: {rotateZ: toggled ? 90 : 0},
+        duration: 300
+    }),
+    drawer: (/* props */) => ({
+        enter: {
+            animation: 'slideDown',
             duration: 300
-        };
-    },
-    drawer: (/* props */) => {
-        return {
-            enter: {
-                animation: 'slideDown',
-                duration: 300
-            },
-            leave: {
-                animation: 'slideUp',
-                duration: 300
-            }
-        };
-    }
+        },
+        leave: {
+            animation: 'slideUp',
+            duration: 300
+        }
+    })
 };
